refactor(stores): split accountStore, rename JSX-free stores to .ts
- extract the security/provisioning builders (provisionAccount,
  buildSecurityFromPassword, buildSecurityFromBiometrics) into
  stores/utils/accountSecurity.ts
- extract the three module-level subscriptions (publish public key,
  provider init, MNS fetch) into accountStore.subscriptions.ts,
  registered explicitly after store creation (type-only circular
  import, TDZ-safe)
- rename the ten JSX-free store files from .tsx to .ts (matching the
  messageStore helpers/events/types split precedent) and fix the one
  explicit-extension import in App.tsx
