+++ src/stores/selfMessageStore.ts
